ForestCMS - Filesystem-Oriented reST CMS
ForestCMS stands for Filesystem-Oriented reStructuredText Content
Management System. It's inspired in the ideas brought by the Hooks PHP
framework for content managements and it's intended to make easy to
publish contents in a filesystem-oriented way (i.e. use the hierarchical
nature of the filesystem to build a tree-like content management). This
makes very easy to manage contents without a web interface (by just
editing files in the filesystem). A web interface can come in the future,
thoug.

ForestCMS is written in Python, and the content is written in reST instead
of HTML. This is the initial commit with just a proof of concept, a small
web server that serve reST files as if they were web pages.